+++ DTO/CartDTO.cs
@@ -4,7 +4,7 @@
 
         public string ProductName { get; set; }
 
-        public int? Quantity { get; set; }
+        public short? Quantity { get; set; }
 
         public string? QuantityPerUnit { get; set; }
